backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
pydantic
passlib[bcrypt]==1.7.4
python-jose
pydantic[email]
bcrypt==4.0.1
pandas
python-multipart
xlrd
openpyxl
cryptography
pycryptodome